use-spring-未测试

+++ rbac/src/main/java/dtx/oa/rbac/dao/RoleDao.java
@@ -5,64 +5,70 @@
  */
 package dtx.oa.rbac.dao;
 
+import dtx.oa.rbac.basic.BasicDao;
 import dtx.oa.rbac.idao.IRoleDao;
 import dtx.oa.rbac.idao.factory.IDaoFactory;
 import dtx.oa.rbac.model.Role;
 import dtx.oa.rbac.model.RoleTree;
-import dtx.oa.util.HibernateUtil;
-import dtx.oa.util.LogUtil;
 import java.util.List;
-import org.hibernate.HibernateException;
-import org.hibernate.Query;
-import org.hibernate.Session;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  *
  * @author datouxi
  */
-public class RoleDao implements IRoleDao {
+@Transactional
+public class RoleDao extends BasicDao implements IRoleDao {
 
     @Override
     public Role getRoleById(String roleId) {
-        Session session=HibernateUtil.getSession();
-        session.beginTransaction();
-        Role role=(Role) session.get(Role.class, roleId);
-        session.getTransaction().commit();
-        return role;
+        return (Role) findById(Role.class, roleId);
+        
+//        Session session=HibernateUtil.getSession();
+//        session.beginTransaction();
+//        Role role=(Role) session.get(Role.class, roleId);
+//        session.getTransaction().commit();
+//        return role;
     }
 
     @Override
     public List<Role> getByStatus(boolean status) {
-        Session session=HibernateUtil.getSession();
-        session.beginTransaction();
-        Query query=session.createQuery("FROM Role role WHERE role.status=:status");
-        query.setBoolean("status", status);
-        List<Role> result=query.list();
-        session.getTransaction().commit();
-        return result;
+        return executeQuery("FROM Role role WHERE role.status=?", new Object[]{status});
+        
+//        Session session=HibernateUtil.getSession();
+//        session.beginTransaction();
+//        Query query=session.createQuery("FROM Role role WHERE role.status=:status");
+//        query.setBoolean("status", status);
+//        List<Role> result=query.list();
+//        session.getTransaction().commit();
+//        return result;
     }
 
     @Override
     public List<Role> getChilds(String parentId) {
-        Session session=HibernateUtil.getSession();
-        session.beginTransaction();
-        Query query=session.createQuery("FROM Role role WHERE role.parentId=:parent_id");
-        query.setString("parent_id", parentId);
-        List<Role> result=query.list();
-        session.getTransaction().commit();
-        return result;
+        return executeQuery("FROM Role role WHERE role.parentId=?", new Object[]{parentId});
+        
+//        Session session=HibernateUtil.getSession();
+//        session.beginTransaction();
+//        Query query=session.createQuery("FROM Role role WHERE role.parentId=:parent_id");
+//        query.setString("parent_id", parentId);
+//        List<Role> result=query.list();
+//        session.getTransaction().commit();
+//        return result;
     }
 
     @Override
     public List<Role> getChilds(String parentId, boolean status) {
-        Session session=HibernateUtil.getSession();
-        session.beginTransaction();
-        Query query=session.createQuery("FROM Role role WHERE role.parentId=:parent_id and role.status=:status");
-        query.setString("parent_id", parentId);
-        query.setBoolean("status", status);
-        List<Role> result=query.list();
-        session.getTransaction().commit();
-        return result;
+        return executeQuery("FROM Role role WHERE role.parentId=? and role.status=?", new Object[]{parentId,status});
+        
+//        Session session=HibernateUtil.getSession();
+//        session.beginTransaction();
+//        Query query=session.createQuery("FROM Role role WHERE role.parentId=:parent_id and role.status=:status");
+//        query.setString("parent_id", parentId);
+//        query.setBoolean("status", status);
+//        List<Role> result=query.list();
+//        session.getTransaction().commit();
+//        return result;
     }
 
     @Override
@@ -87,54 +93,62 @@ public class RoleDao implements IRoleDao {
 
     @Override
     public boolean updateRole(Role role) {
-        boolean result=true;
-        Session session=HibernateUtil.getSession();
-        session.beginTransaction();
-        try{
-            session.update(role);
-        }catch(HibernateException e){
-            LogUtil.getLogger().error(e);
-            result=false;
-        }
-        session.getTransaction().commit();
-        return result;
+        return update(role);
+        
+//        boolean result=true;
+//        Session session=HibernateUtil.getSession();
+//        session.beginTransaction();
+//        try{
+//            session.update(role);
+//        }catch(HibernateException e){
+//            LogUtil.getLogger().error(e);
+//            result=false;
+//        }
+//        session.getTransaction().commit();
+//        return result;
     }
 
     @Override
     public boolean updateRoleMessage(Role role) {
-        Session session=HibernateUtil.getSession();
-        session.beginTransaction();
-        Query query=session.createQuery("UPDATE Role role SET role.roleName=:role_name,role.remark=:remark WHERE role.uuid=:id");
-        query.setString("role_name", role.getRoleName());
-        query.setString("remark", role.getRemark());
-        query.setString("id", role.getUuid());
-        int result=query.executeUpdate();
-        session.getTransaction().commit();
-        return result>0;
+        return update("UPDATE Role role SET role.roleName=?,role.remark=? WHERE role.uuid=?", new Object[]{role.getRoleName(),role.getRemark(),role.getUuid()})>0;
+        
+//        Session session=HibernateUtil.getSession();
+//        session.beginTransaction();
+//        Query query=session.createQuery("UPDATE Role role SET role.roleName=:role_name,role.remark=:remark WHERE role.uuid=:id");
+//        query.setString("role_name", role.getRoleName());
+//        query.setString("remark", role.getRemark());
+//        query.setString("id", role.getUuid());
+//        int result=query.executeUpdate();
+//        session.getTransaction().commit();
+//        return result>0;
     }
 
     @Override
     public boolean updateParent(Role role) {
-        Session session=HibernateUtil.getSession();
-        session.beginTransaction();
-        Query query=session.createQuery("UPDATE Role role SET role.parentId=:parent_id WHERE role.uuid=:id");
-        query.setString("parent_id", role.getParentId());
-        query.setString("id", role.getUuid());
-        int result=query.executeUpdate();
-        session.getTransaction().commit();
-        return result>0;
+        return update("UPDATE Role role SET role.parentId=? WHERE role.uuid=?", new Object[]{role.getParentId(),role.getUuid()})>0;
+        
+//        Session session=HibernateUtil.getSession();
+//        session.beginTransaction();
+//        Query query=session.createQuery("UPDATE Role role SET role.parentId=:parent_id WHERE role.uuid=:id");
+//        query.setString("parent_id", role.getParentId());
+//        query.setString("id", role.getUuid());
+//        int result=query.executeUpdate();
+//        session.getTransaction().commit();
+//        return result>0;
     }
 
     @Override
     public boolean updateStatus(Role role) {
-        Session session=HibernateUtil.getSession();
-        session.beginTransaction();
-        Query query=session.createQuery("UPDATE Role role SET role.status=:status WHERE role.uuid=:id");
-        query.setBoolean("status", role.getStatus());
-        query.setString("id", role.getUuid());
-        int result=query.executeUpdate();
-        session.getTransaction().commit();
-        return result>0;
+        return update("UPDATE Role role SET role.status=? WHERE role.uuid=?", new Object[]{role.getStatus(),role.getUuid()})>0;
+        
+//        Session session=HibernateUtil.getSession();
+//        session.beginTransaction();
+//        Query query=session.createQuery("UPDATE Role role SET role.status=:status WHERE role.uuid=:id");
+//        query.setBoolean("status", role.getStatus());
+//        query.setString("id", role.getUuid());
+//        int result=query.executeUpdate();
+//        session.getTransaction().commit();
+//        return result>0;
     }
 
     @Override
@@ -144,23 +158,28 @@ public class RoleDao implements IRoleDao {
 
     @Override
     public boolean deleteRole(String id) {
-        Session session=HibernateUtil.getSession();
-        session.beginTransaction();
-        Query query=session.createQuery("DELETE FROM Role role WHERE role.uuid=:id");
-        query.setString("id", id);
-        int result=query.executeUpdate();
-        session.getTransaction().commit();
-        if(result<=0)return false;
-        return IDaoFactory.iRoleNodeDao().deleteByRoleId(id)&&IDaoFactory.iRoleUserDao().deleteByRoleId(id);
+        if(update("DELETE FROM Role role WHERE role.uuid=?", new Object[]{id})>0)return IDaoFactory.iRoleNodeDao().deleteByRoleId(id)&&IDaoFactory.iRoleUserDao().deleteByRoleId(id);
+        return false;
+        
+//        Session session=HibernateUtil.getSession();
+//        session.beginTransaction();
+//        Query query=session.createQuery("DELETE FROM Role role WHERE role.uuid=:id");
+//        query.setString("id", id);
+//        int result=query.executeUpdate();
+//        session.getTransaction().commit();
+//        if(result<=0)return false;
+//        return IDaoFactory.iRoleNodeDao().deleteByRoleId(id)&&IDaoFactory.iRoleUserDao().deleteByRoleId(id);
     }
 
     @Override
     public String addRole(Role role) {
-        Session session=HibernateUtil.getSession();
-        session.beginTransaction();
-        String result=(String) session.save(role);
-        session.getTransaction().commit();
-        return result;
+        return (String) add(role);
+        
+//        Session session=HibernateUtil.getSession();
+//        session.beginTransaction();
+//        String result=(String) session.save(role);
+//        session.getTransaction().commit();
+//        return result;
     }
     
 }
